Chore(.user): update package
