+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/KakaoAuthorizeFeignClient.java
@@ -4,8 +4,8 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import com.newbarams.ajaja.module.user.auth.model.KakaoResponse;
-import com.newbarams.ajaja.module.user.auth.model.KakaoTokenRequest;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoTokenRequest;
 
 @FeignClient(name = "KakaoAuthorizeFeignClient", url = "https://kauth.kakao.com")
 public interface KakaoAuthorizeFeignClient {
