+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/KakaoProfileFeignClient.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import com.newbarams.ajaja.module.user.auth.model.KakaoResponse;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
 
 @FeignClient(name = "KakaoProfileFeignClient", url = "https://kapi.kakao.com")
 public interface KakaoProfileFeignClient {
