+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/KakaoProperties.java
@@ -10,4 +10,6 @@ import lombok.RequiredArgsConstructor;
 @ConfigurationProperties(prefix = "secret.kakao")
 public class KakaoProperties {
 	private final String clientId;
+	private final String clientSecret;
+	private final String logoutRedirectUrl;
 }
