+++ src/main/java/com/newbarams/ajaja/module/user/application/AuthorizeService.java
@@ -2,7 +2,7 @@ package com.newbarams.ajaja.module.user.application;
 
 import org.springframework.stereotype.Service;
 
-import com.newbarams.ajaja.module.user.auth.model.AccessToken;
+import com.newbarams.ajaja.module.user.application.model.AccessToken;
 
 @Service
 public interface AuthorizeService {
