--- src/main/java/com/newbarams/ajaja/module/user/application/EmailVerification.java
@@ -1,7 +1,0 @@
-package com.newbarams.ajaja.module.user.application;
-
-record EmailVerification(
-	String email,
-	String certification
-) {
-}
