+++ src/main/java/com/newbarams/ajaja/module/user/application/SendVerificationEmailService.java
@@ -6,6 +6,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.newbarams.ajaja.module.user.application.model.EmailVerification;
 import com.newbarams.ajaja.module.user.domain.User;
 
 import lombok.RequiredArgsConstructor;
