+++ src/main/java/com/newbarams/ajaja/module/user/application/VerifyCertificationService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.newbarams.ajaja.global.common.exception.AjajaException;
+import com.newbarams.ajaja.module.user.application.model.EmailVerification;
 import com.newbarams.ajaja.module.user.domain.User;
 
 import lombok.RequiredArgsConstructor;
