+++ src/main/java/com/newbarams/ajaja/module/user/application/model/EmailVerification.java
@@ -1,0 +1,7 @@
+package com.newbarams.ajaja.module.user.application.model;
+
+public record EmailVerification(
+	String email,
+	String certification
+) {
+}
