--- src/main/java/com/newbarams/ajaja/module/user/auth/model/AccessToken.java
@@ -1,5 +1,0 @@
-package com.newbarams.ajaja.module.user.auth.model;
-
-public interface AccessToken {
-	String getContent();
-}
