+++ src/main/java/com/newbarams/ajaja/module/user/kakao/application/KakaoGetProfileService.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.module.user.auth.application;
+package com.newbarams.ajaja.module.user.kakao.application;
 
 import java.util.Objects;
 
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 import com.newbarams.ajaja.infra.feign.kakao.KakaoProfileFeignClient;
 import com.newbarams.ajaja.module.user.application.GetProfileService;
-import com.newbarams.ajaja.module.user.auth.model.Profile;
+import com.newbarams.ajaja.module.user.application.model.Profile;
 
 import lombok.RequiredArgsConstructor;
