+++ src/main/java/com/newbarams/ajaja/module/user/kakao/model/KakaoResponse.java
@@ -1,10 +1,12 @@
-package com.newbarams.ajaja.module.user.auth.model;
+package com.newbarams.ajaja.module.user.kakao.model;
 
-import static com.newbarams.ajaja.module.user.auth.model.KakaoResponse.*;
+import static com.newbarams.ajaja.module.user.kakao.model.KakaoResponse.*;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import com.newbarams.ajaja.module.user.application.model.AccessToken;
+import com.newbarams.ajaja.module.user.application.model.Profile;
 
 public sealed interface KakaoResponse permits Token, UserInfo {
 	@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
