+++ src/main/resources/application-secret.yml
@@ -5,4 +5,5 @@ secret:
 
   kakao:
     client-id: ENC(egCrmTnMY/kGZNYJwP7O+dHISas8fVWFmO4bF9M6zkDHi02moIaexhENauQDBduD)
-    client-secret: ENC(g/sAYLisOXOhvq7eFD2z4kwI4MkHbJQQBJGXDoEOIEq87p7zZdYWMNy+OPJwb56T)
+    client-secret: ENC(Eo3sbkxTa8NAnpb3OOhhhKMeu+O9KBpsIQsDL57wiISyA6T+wqfaaeJglbyurx3h)
+    logout-redirect-url: ENC(cQFggzXPK6NQs7Aj/K1V5zlkGvJcXLtvNPsWBx7KiQ2DFxeoMi2+w41q9go7gIoHQ1pDFM+l++g=)
