+++ src/test/java/com/newbarams/ajaja/module/user/application/LoginServiceTest.java
@@ -13,12 +13,12 @@ import org.mockito.Mock;
 
 import com.newbarams.ajaja.common.MockTestSupport;
 import com.newbarams.ajaja.global.security.jwt.util.JwtGenerator;
-import com.newbarams.ajaja.module.user.auth.model.AccessToken;
-import com.newbarams.ajaja.module.user.auth.model.KakaoAccount;
-import com.newbarams.ajaja.module.user.auth.model.KakaoResponse;
-import com.newbarams.ajaja.module.user.auth.model.Profile;
+import com.newbarams.ajaja.module.user.application.model.AccessToken;
+import com.newbarams.ajaja.module.user.application.model.Profile;
 import com.newbarams.ajaja.module.user.domain.User;
 import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoAccount;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
 
 class LoginServiceTest extends MockTestSupport {
 	@InjectMocks
