+++ src/test/java/com/newbarams/ajaja/module/user/application/VerifyCertificationServiceTest.java
@@ -12,6 +12,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import com.newbarams.ajaja.common.MonkeySupport;
 import com.newbarams.ajaja.common.RedisBasedTest;
 import com.newbarams.ajaja.global.common.exception.AjajaException;
+import com.newbarams.ajaja.module.user.application.model.EmailVerification;
 import com.newbarams.ajaja.module.user.domain.Email;
 import com.newbarams.ajaja.module.user.domain.User;
 import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
