+++ src/test/java/com/newbarams/ajaja/module/user/kakao/application/KakaoGetProfileServiceTest.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.module.user.auth.application;
+package com.newbarams.ajaja.module.user.kakao.application;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
@@ -10,8 +10,8 @@ import org.mockito.Mock;
 
 import com.newbarams.ajaja.common.MockTestSupport;
 import com.newbarams.ajaja.infra.feign.kakao.KakaoProfileFeignClient;
-import com.newbarams.ajaja.module.user.auth.model.KakaoResponse;
-import com.newbarams.ajaja.module.user.auth.model.Profile;
+import com.newbarams.ajaja.module.user.application.model.Profile;
+import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
 
 class KakaoGetProfileServiceTest extends MockTestSupport {
 	@InjectMocks
